requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.3
black==23.12.0
bracex==2.4
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chatbotAI==0.3.1.3
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
distro==1.9.0
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
icecream==2.1.3
idna==3.6
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
lxml==5.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
openai==1.21.1
packaging==23.2
parameterized==0.9.0
pathspec==0.12.1
platformdirs==4.1.0
pvleopard==2.0.2
pvporcupine==3.0.2
pvrecorder==1.2.2
PyAudio==0.2.14
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
python-docx==1.1.0
python-dotenv==1.0.1
pyttsx3==2.90
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
resolvelib==1.0.1
rich==13.7.0
rpds-py==0.13.2
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.3
srt==3.5.3
subprocess-tee==0.4.1
tabulate==0.9.0
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.1.0
vosk==0.3.45
wcmatch==8.5
websockets==12.0
wheel==0.41.2
wikipedia==1.4.0
yamllint==1.33.0
yarl==1.9.4
